feat: mediator rename package and code cleanup
