+++ mediator/src/main/resources/application.conf
