+++ mediator/src/main/scala/io/iohk/atala/mediator/actions/ProtocolExecute.scala
@@ -1,18 +1,17 @@
-package fmgp.did.comm.protocol
+package io.iohk.atala.mediator.actions
 
-import zio._
-import zio.json._
-
-import fmgp.crypto.error._
-import fmgp.did._
-import fmgp.did.comm._
-import fmgp.did.comm.Operations._
-import fmgp.did.comm.protocol._
-import fmgp.did.comm.protocol.basicmessage2._
-import fmgp.did.comm.protocol.trustping2._
-import fmgp.did.db._
+import fmgp.crypto.error.*
+import fmgp.did.*
+import fmgp.did.comm.*
+import fmgp.did.comm.Operations.*
+import fmgp.did.comm.protocol.*
+import fmgp.did.comm.protocol.basicmessage2.*
+import fmgp.did.comm.protocol.trustping2.*
+import io.iohk.atala.mediator.*
 import io.iohk.atala.mediator.comm.*
-
+import io.iohk.atala.mediator.db.*
+import zio.*
+import zio.json.*
 //TODO pick a better name // maybe "Protocol" only
 
 trait ProtocolExecuter[-R] {
