+++ mediator/src/main/scala/io/iohk/atala/mediator/protocols/ForwardMessageExecuter.scala
@@ -1,13 +1,15 @@
-package fmgp.did.comm.mediator
+package io.iohk.atala.mediator.protocols
 
-import zio._
-import zio.json._
-import fmgp.crypto.error._
-import fmgp.did._
-import fmgp.did.comm._
-import fmgp.did.comm.protocol._
-import fmgp.did.comm.protocol.routing2._
-import fmgp.did.db._
+import fmgp.crypto.error.*
+import fmgp.did.*
+import fmgp.did.comm.*
+import fmgp.did.comm.protocol.*
+import fmgp.did.comm.protocol.routing2.*
+import io.iohk.atala.mediator.MediatorError
+import io.iohk.atala.mediator.actions.*
+import io.iohk.atala.mediator.db.*
+import zio.*
+import zio.json.*
 
 object ForwardMessageExecuter
     extends ProtocolExecuterWithServices[
