+++ mediator/src/main/scala/io/iohk/atala/mediator/protocols/PickupExecuter.scala
@@ -1,15 +1,16 @@
-package fmgp.did.comm.mediator
+package io.iohk.atala.mediator.protocols
 
-import zio._
-import zio.json._
-import fmgp.crypto.error._
-import fmgp.did._
-import fmgp.did.comm._
-import fmgp.did.comm.Operations._
-import fmgp.did.comm.protocol._
-import fmgp.did.comm.protocol.pickup3._
-import fmgp.did.db._
-
+import fmgp.crypto.error.*
+import fmgp.did.*
+import fmgp.did.comm.*
+import fmgp.did.comm.Operations.*
+import fmgp.did.comm.protocol.*
+import fmgp.did.comm.protocol.pickup3.*
+import io.iohk.atala.mediator.*
+import io.iohk.atala.mediator.actions.*
+import io.iohk.atala.mediator.db.*
+import zio.*
+import zio.json.*
 object PickupExecuter
     extends ProtocolExecuterWithServices[ProtocolExecuter.Services & DidAccountRepo & MessageItemRepo] {
